% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_functions.R
\name{boxplot_pmt}
\alias{boxplot_pmt}
\title{Make A Boxplot of Payments By DRG Code}
\usage{
boxplot_pmt(pmt)
}
\arguments{
\item{pmt}{A payment category in  Medicare Severity Diagnosis Related Group data (DRG data)}
}
\value{
A boxplot of payments by DRG code
}
\description{
Make A Boxplot of Payments By DRG Code
}
\examples{
boxplot_pmt('Average_Covered_Charges')

}
