% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_functions.R
\name{makeBox}
\alias{makeBox}
\title{Calculate A Statistics of Average Medicare Payments}
\usage{
makeBox(option)
}
\arguments{
\item{option}{A statistics (mean, median, or standard deviation) that is required to be calculated}
}
\value{
A mean, median, or standard deviation of average Medicare payments
}
\description{
Calculate A Statistics of Average Medicare Payments
}
\examples{
makeBox("Mean")
}
